refactor: Move diff_version_service.dart to services folder
